Improve handling of NMEA messages
0b2aeb8 added endings to each message, but the network client already
splits on newlines. Additionally, Yacht Devices gateways return two
newlines between messages. It's unclear whether that's the device
doing that or if the messages on the bus itself have that ending.
Either way, this makes the processing a bit more robust. As long as
there is at least one newline between messages, this should work.